+++ picows/__init__.py
@@ -1,9 +1,16 @@
 from .types import (
     WSError,
+    WSUpgradeFailure,
+    WSProtocolError,
     WSUpgradeRequest,
     WSUpgradeResponse,
     WSUpgradeResponseWithListener,
     PICOWS_DEBUG_LL
+)
+
+from .url import (
+    WSInvalidURL,
+    WSParsedURL
 )
 
 from .picows import (
@@ -15,30 +22,30 @@ from .picows import (
     WSListener,
 )
 
-from .url import (
-    WSInvalidURL
-)
-
 from .api import (
     ws_connect,
-    ws_create_server
+    ws_create_server,
 )
 
 __all__ = [
     'WSError',
+    'WSUpgradeFailure',
+    'WSProtocolError',
+    'WSUpgradeRequest',
+    'WSUpgradeResponse',
+    'WSUpgradeResponseWithListener',
+    'PICOWS_DEBUG_LL',
+    'WSInvalidURL',
+    'WSParsedURL',
     'WSMsgType',
     'WSCloseCode',
     'WSAutoPingStrategy',
     'WSFrame',
     'WSTransport',
     'WSListener',
-    'WSUpgradeRequest',
-    'WSUpgradeResponse',
-    'WSUpgradeResponseWithListener',
     'ws_connect',
     'ws_create_server',
-    'PICOWS_DEBUG_LL'
 ]
 
-__version__ = "1.15.0"
+__version__ = "1.16.0"
 __author__ = "Taras Kozlov"